Replace feed edit/rules sheets with NavigationStack push views
The .sheet(item:) + @Binding to optional Feed pattern triggered a
SwiftUI bug where sheet @State persisted across presentations, leaking
the previous feed's data into a freshly opened sheet. Renames the views
to FeedEditView and FeedRulesView, takes the feed as a non-optional let,
initializes @State from init(), uses Environment(.dismiss), and switches
both call sites to navigationDestination(item:). Also drops the
feedToEdit nil-then-async workaround in iPadSidebarView.
